requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
Pillow>=9.0.0
scipy>=1.7.0
shap>=0.41.0
PyYAML>=6.0
